Feature/pp9 9012 shares

+++ src/picturepark-sdk-v1-angular/src/app/share-manager/components/share-manager-item/share-manager-item.component.ts
@@ -4,7 +4,7 @@ import { Subscription } from 'rxjs';
 
 // LIBRARIES
 import { ShareService, ShareContentDetail, IShareDataBasic, IMailRecipient, ShareDeleteManyRequest, ShareDetail } from '@picturepark/sdk-v1-angular';
-import { ContentDownloadDialogService } from '@picturepark/sdk-v1-angular-ui';
+import { ContentDownloadDialogService, DialogService, TranslationService } from '@picturepark/sdk-v1-angular-ui';
 
 @Component({
   selector: 'app-share-manager-item',
@@ -33,6 +33,8 @@ export class ShareManagerItemComponent implements OnInit, OnDestroy {
     private activatedRoute: ActivatedRoute,
     private shareService: ShareService,
     private contentDownloadService: ContentDownloadDialogService,
+    private dialogService: DialogService,
+    private translationService: TranslationService,
     private router: Router
   ) {
 
@@ -63,10 +65,21 @@ export class ShareManagerItemComponent implements OnInit, OnDestroy {
       name: 'Delete',
       icon: 'delete',
       action: () => {
-        // TODO: Add confirm dialog
-        this.shareService.deleteMany(new ShareDeleteManyRequest({ ids: [this.share.id] }))
-        .subscribe(i => {
-          this.router.navigate(['./share-manager']);
+        this.dialogService.confirm({
+          title: this.translationService.translate('ShareManager.DeleteShare'),
+          message: this.translationService.translate('ShareManager.ConfirmDelete'),
+          options: {
+            okText: this.translationService.translate('ShareManager.Delete'),
+            cancelText: this.translationService.translate('ShareManager.Cancel')
+          }
+        })
+        .afterClosed().subscribe(result => {
+          if (result) {
+            this.shareService.deleteMany(new ShareDeleteManyRequest({ ids: [this.share.id] }))
+            .subscribe(i => {
+              this.router.navigate(['./share-manager']);
+            });
+          }
         });
       }
     }];
